refactor(Calculator): 숫자 리스트 합 반환 타입 int로 수정

+++ src/main/java/calculator/Application.java
@@ -6,7 +6,7 @@ public class Application {
         Calculator calculator = new SumCalculator();
         OutputView outputView = new OutputView();
 
-        long result = calculator.sum(userInput.createFormula());
+        int result = calculator.sum(userInput.createFormula());
         outputView.getSumResult(result);
     }
 }
